+++ CHANGELOG.md
@@ -10,3 +10,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   transform items, place items, and handling failures based on previous steps.
 - Basic rabbit-mq node with a bear-bones implementation.
 - Started basic high level docs.
+- Added MMOP/0 worker-ready.
+- Added MMOP/0 start-node.